Moved Interfaces and Types to FPLite.cs, Created abstract class for UnwrapException
