--- FPLite/Union/UnionType.cs
@@ -1,13 +1,0 @@
-﻿namespace FPLite.Union;
-
-public enum UnionType : byte
-{
-    T1,
-    T2,
-    T3,
-    T4,
-    T5,
-    T6,
-    T7,
-    T8
-}
